requirements.txt
flask>=3.0.0
ddddocr>=1.5.6
pillow>=10.0.0
